More endpoint moving
Getting close to narrowing down to just the endpoints that require access from both the instrument server and frontend
